feat: add branding preview component and integrate into branding settings
